Refactor jDKIM to return all signatures results
This is needed to properly handle RFC 7001 and RFC 6591.
Even if one signature is in error, we still need *all* signature results
in order to decide what to log, do and report.

Instead of subclassing Exceptions, require a failure reason in
constructors.
Reasons are enumerated, and a textual explanation (which may be logged
and included in reports) is accessible for each reason.