Enforce absolute domain for DNS lookups
DNS queries do distinguish relative and absolute domain names.
If the domain does not end with a dot '.' (which is what we deal
with usually in SMTP values) and the query does not return anything
(or fails) the resolver may try again by suffixing the configured
system (resolv.conf) DNS search domains.
Since we actually wish to query only the domain name we target,
ensure to use its absolute form for lookups.